refactor to support txs with 4 byte version field #27

+++ src/infrastructure/schemas/AddressAliasTransactionSchema.ts
@@ -34,7 +34,7 @@ export default new Schema([
     uint('size'),
     array('signature'),
     array('signer'),
-    ushort('version'),
+    uint('version'),
     ushort('type'),
     array('fee', TypeSize.INT),
     array('deadline', TypeSize.INT),
